requirements.txt
streamlit==1.17.0
os
uuid
base64
json
speechrecognition==3.9.0
gTTS==2.2.4
audio-recorder-streamlit==0.1.2
streamlit-float==0.0.1
python-dotenv==0.21.0
groq==0.1.0
librosa==0.10.0
soundfile==0.10.3.post1
langdetect==1.0.9
elevenlabs==0.1.3
torch==2.1.0+cpu
